requirements.txt
jupyter
ipykernel
pandas
geopandas
matplotlib
seaborn
scikit-learn
shapely
mapclassify
folium
contextily
scipy
pymongo
python-dotenv
openpyxl
streamlit>=1.32.0
